Functionality/events (#3)
* Started working on delegates

Note: Delegate::Remove doesn't work right now. Maybe comparing the target of two std::functions could work. (But does that work with raw function pointers? And how would I know the type to look for if target<DelegateArgs...>() doesn't work?)

* Unsubscribing now works

(Now storing a pointer to the subscribed object as well)

* Added documentation

* Added macros to subscribe to delegates

* Added support for static function subscription in delegates

Also, Add now checks if a function is valid when it's being added

* Made some changes towards moving Delegate to use any class rather than a list of args

* Renamed Subscribe_OneParam macro to Subscribe, as this is likely used more often

Also renamed the class to EventBase rather than DelegateBase, as the class sends out events to delegates, and is not a delegate itself.